Time: 1 ms (100%), Space: 49.1 MB (15.85%) - LeetHub

+++ 0004-median-of-two-sorted-arrays/0004-median-of-two-sorted-arrays.java
@@ -32,7 +32,7 @@ class Solution {
             mean=(double)arr[n/2];
         }
         else{
-           double sum=arr[(n/2)-1]+arr[(n/2)];
+           double sum=(double)arr[(n/2)-1]+(double)arr[(n/2)];
             mean=sum/2;
         }
         return mean;
